gacha-gen: more banner image gen stuff
- improve text styling (title and description - colours, spacing, etc.)
- support split backgrounds
- better card randomisation (limit to one of each chara unless one-chara series)
- add some padding around quotes in description text
- split bulk of image gen to separate module so it's out of the way

minor gacha data adjustment too:
- added a line break to chinese and wedding (week 13) text
