+++ gacha_data/limited_rotation.csv
@@ -13,7 +13,8 @@ ID Ranges,755-758,"804-805, 806-807",248-250,242-244,255-256,,,"750-751, 765-767
 Banner Text (ja),「巫女さん」シリーズ,「迎春」&「初詣」シリーズ,「謹賀新年」シリーズ,「ゆく年くる年」シリーズ,「お正月」シリーズ,,,「冬キャンプ」シリーズ,「節分」シリーズ,"ハニプラ & B.A.C
 「バレンタイン」シリーズ","ハニプラ & 2_wEi
 「バレンタイン」シリーズ",「初恋」シリーズ,"B.A.C
-「Innocence」シリーズ",「ひな祭り」シリーズ,「Precious」シリーズ,「Precious」シリーズ,「ウェディング」シリーズ,「中華」&「ウェディング」シリーズ,「イースター」シリーズ,,「夢の世界」シリーズ,"ハニプラ「雨宿り」&
+「Innocence」シリーズ",「ひな祭り」シリーズ,「Precious」シリーズ,「Precious」シリーズ,「ウェディング」シリーズ,"「中華」&
+「ウェディング」シリーズ",「イースター」シリーズ,,「夢の世界」シリーズ,"ハニプラ「雨宿り」&
 2_wEi「梅雨」シリーズ","「マジシャン」シリーズ
 &「サーカス」ひなた",「パジャマ」シリーズ,「料理」シリーズ,「白衣」シリーズ,「妖精」シリーズ,"「妖精」彩芽 &
 「人魚」ほたる",,「サマーバカンス」シリーズ,"ハニプラ「夏休み」&
@@ -34,7 +35,8 @@ Banner Text (en),"""Shrine Maiden"" Series","""New Year's Greetings"" &
 ""Hatsumoude"" Series","""Happy New Year"" Series","""Start of the New Year"" Series","""New Year's Day"" Series",,,"""Winter Camping"" Series","""Setsubun"" Series","8/pLanet!! & B.A.C
 ""Valentine's"" Series","8/pLanet!! & 2_wEi
 ""Valentine's"" Series","""First Love"" Series","B.A.C
-""Innocence"" Series","""Dolls Day"" Series","""Precious"" Series","""Precious"" Series","""Wedding"" Series","""Chinese"" & ""Wedding"" Series","""Easter"" Series",,"""Dreamland"" Series","8/pLanet!! ""Shelter from the Rain""
+""Innocence"" Series","""Dolls Day"" Series","""Precious"" Series","""Precious"" Series","""Wedding"" Series","""Chinese"" &
+""Wedding"" Series","""Easter"" Series",,"""Dreamland"" Series","8/pLanet!! ""Shelter from the Rain""
 & 2_wEi ""Rainy Season"" Series","""Magician"" Series
 & ""Circus"" Hinata","""Pajamas"" Series","""Cooking"" Series","""White Coat"" Series","""Fairy"" Series","""Fairy"" Ayame &
 ""Mermaid"" Hotaru",,"""Summer Vacation"" Series","8/pLanet!! ""Summer Break"" &
